#include <X11/extensions/Xrandr.h> if we have RandR support, for
Mon Oct 27 15:47:58 2003  Manish Singh  <yosh@gimp.org>

        * gdk/x11/gdkscreen-x11.c: #include <X11/extensions/Xrandr.h> if
        we have RandR support, for XRRUpdateConfiguration declaration.

        * gtk/gtkfilefilter.c: #include "gtkprivate.h" for _gtk_fnmatch
        declaration.

        * gtk/gtkuimanager.c (gtk_ui_manager_add_ui_from_file): Make
        length a gsize variable, since that's what g_file_get_contents()
        expects.

+++ gtk/gtkuimanager.c
@@ -1280,7 +1280,7 @@ gtk_ui_manager_add_ui_from_file (GtkUIManager *self,
 				 GError      **error)
 {
   gchar *buffer;
-  gint length;
+  gsize length;
   guint res;
 
   g_return_val_if_fail (GTK_IS_UI_MANAGER (self), 0);
